Implement CreateBackupAction.perform with script

+++ lib/backmeup/create_backup_action.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'tty-command'
+
 module Backmeup
   # Create a backup
   class CreateBackupAction
@@ -16,8 +18,12 @@ module Backmeup
     attr_reader :destination, :previous_destination, :root
 
     def perform
-      File.open(File.join(root.config, 'filelist')).each do |src|
-        FileUtils.cp_r(src.chomp, backup_destination)
+      if script_exists?
+        cmd.run(script_pathname.to_s)
+      else
+        File.open(File.join(root.config, 'filelist')).each do |src|
+          FileUtils.cp_r(src.chomp, backup_destination)
+        end
       end
     end
 
@@ -25,6 +31,27 @@ module Backmeup
 
     def backup_destination
       @backup_destination ||= File.join(root.backups, destination, 'data')
+    end
+
+    def cmd
+      TTY::Command.new
+    end
+
+    def script_exists?
+      root.bin.glob(script_name)[0] ? true : false
+    end
+
+    def script_path
+      script_pathname.to_s
+    end
+
+    def script_pathname
+      @script_pathname ||= Pathname.new(File.join(root.bin, script_name))
+    end
+
+    def script_name
+      match = self.class.to_s.match(/\A.*::(.+)Action\z/) || return
+      match[1].underscore
     end
   end
 end
